Add files via upload
"This project applies CNN image segmentation for salt identification using the U-Net model. The data for training and testing was obtained from the TGS-Kaggle Salt Identification challenge in 2018 (https://www.kaggle.com/c/tgs-salt-identification-challenge/). The training dataset consists of 4000 (101 by 101) seismic patches, along with their corresponding masks that indicate the presence or absence of salt. The test dataset consists of 18000 (101 by 101) seismic patches. The goal of the project is to predict the regions of salt on the test seismic patches using the U-Net model."